project - Copy/requirements.txt
Flask
google-generativeai
python-dotenv
gunicorn